Fuse the scripts of run_distributed_engines.py and run_distributed_on_slurm.py
